Working on pathfinding
tried to implement a dijskra map for player position that all the
pirates in range could use for pathfinding. this seemed more efficient
than having each pirate perform an A* each frame. but this map still
needs to be generated every time the player moves to a new tile, and it
is using too much CPU time and memory. We will need to limit the range
of the map to maybe 16 tiles in each direction from player or something.