Setup dijkstra map so that it only scans a certain window
if pirates are within the map window, they will generate paths to
player. once they are no longer within the window, the paths will not be
generated.

next we need to detect when a pirate is within the window but cannot
find a valid path and avoid stack overflow error